vehicles store and loadings
