+++ app/(app)/tickets/filters/vehicles.tsx
@@ -9,15 +9,15 @@ import Divider from '@/components/shared/Divider'
 import PressableStyled from '@/components/shared/PressableStyled'
 import Typography from '@/components/shared/Typography'
 import { useTranslation } from '@/hooks/useTranslation'
-import { useVehicles } from '@/hooks/useVehicles'
 import { VehicleDto } from '@/modules/backend/openapi-generated'
 import { useTicketsFiltersStoreContext } from '@/state/TicketsFiltersStoreProvider/useTicketsFiltersStoreContext'
 import { useTicketsFiltersStoreUpdateContext } from '@/state/TicketsFiltersStoreProvider/useTicketsFiltersStoreUpdateContext'
+import { useVehiclesStoreContext } from '@/state/VehiclesStoreProvider/useVehiclesStoreContext'
 
 const TicketsFiltersVehiclesScreen = () => {
   const t = useTranslation('TicketsFilters')
 
-  const { vehicles } = useVehicles()
+  const { vehicles } = useVehiclesStoreContext()
 
   const onPurchaseStoreUpdate = useTicketsFiltersStoreUpdateContext()
   const { ecv } = useTicketsFiltersStoreContext()
